pyproject.toml
[project]
name = "politeai"
version = "1.0.0"
description = "Korean tone/politeness transformation service"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "openai>=1.58.0",
    "resend>=2.5.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "email-validator>=2.2.0",
    "sse-starlette>=2.2.0",
    "regex>=2024.11.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
